requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
python-dotenv==1.0.0
openai==1.3.3
langchain==0.0.335
psycopg2-binary==2.9.7
bcrypt==4.0.1
python-dateutil==2.8.2 